+++ lib/src/com/hydrologis/dart_shp/shapefile_reader.dart
@@ -157,7 +157,7 @@ class ShapefileReader {
   IndexFile shxReader;
   GeometryFactory geometryFactory;
 
-  bool flatGeometry;
+  bool flatGeometry = false;
 
   bool onlyRandomAccess;