+++ lib/src/com/hydrologis/dart_shp/utils.dart
@@ -5,11 +5,12 @@ final LOGGER = Logger(level: Level.verbose);
 const NULL_CHAR = '\x00';
 final NULL_CHARACTERS = Characters(NULL_CHAR);
 
+// TODO NOT YET IMPLEMENTED
 class TimeZones {
   String _timeZoneName;
 
   static void init() {
-    tz.initializeTimeZones();
+    // tz.initializeTimeZones();
   }
 
   void setZone(String timeZoneName) {
@@ -25,6 +26,7 @@ class TimeZones {
   }
 }
 
+// TODO this just does utf8 right now
 class Charset {
   final String charsetName;